requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
werkzeug==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.1
yfinance==0.2.40
flask-socketio==5.3.6
flask-limiter==3.7.0
redis==5.0.4
celery==5.4.0
statsmodels==0.14.2
python-dotenv==1.0.1
async-timeout>=4.0.2
eventlet==0.35.2
gunicorn==22.0.0
